LayoutTests/http/tests/security/xssAuditor/full-block-script-tag-expected.txt

CONSOLE MESSAGE: line 4: Refused to execute a JavaScript script. Source code of script found within request.

ALERT: Referrer is ""
There should be no content in the iframe below:



--------
Frame: 'frame'
--------
